# Reduce memory usage in `cupy.sort`